new-backend/requirements.txt
# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# Database
asyncpg==0.29.0
alembic==1.13.3
psycopg2-binary==2.9.9  # For Alembic migrations

# Redis
redis==5.1.1
hiredis==3.0.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.0
cryptography==43.0.1

# HTTP Client
httpx[http2]==0.27.2

# Browser Automation
playwright==1.48.0

# Docker SDK (for WAHA orchestration)
docker==7.1.0

# Google Gemini Integration
google-generativeai==0.8.3

# Payment Integration (TipTopPay)
requests==2.32.3

# WhatsApp (WAHA - deployed separately)
# No Python package needed - HTTP API client via httpx

# Validation
pydantic==2.9.2
pydantic-settings==2.6.0
email-validator==2.2.0

# Utilities
python-dotenv==1.0.1
pytz==2024.2

# PDF Processing (для склейки накладных)
pypdf==5.1.0

# Development (removed for production build)
# pytest==8.3.3
# pytest-asyncio==0.24.0
# pytest-cov==6.0.0
# httpx-mock==0.17.0
# black==24.10.0
# ruff==0.7.4
